Move normalizeFields to field-types
